backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.34.0
sqlalchemy>=2.0.36
pydantic>=2.10.0
pydantic-settings>=2.7.0
anthropic>=0.43.0
python-dotenv>=1.0.1
python-multipart>=0.0.20
slowapi>=0.1.9
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
bcrypt==4.0.1
pytest>=8.0.0
httpx>=0.27.0
